Feat: Improved performance by avoiding iterating over the whole Set of NotifierKeyRefs and re-introduced the synchronized block around getReferences()
The Set of NotifierKeyRefs is now a map, where the key is the location of the real file for the NotifierKeyRef. The Values in the Map are a Set that contains all symlinks
that point to that specific file. If a notification about a file change occurs, an event is created for both the location of the real file and the symlink.